Allow matching strings to natural types
This makes sense because data sent to the server is deserialized - meaning it
might be the value or it could be a string representation of the value.  This
provides a sensible and consistent guessing mechanism that treats strings as
the values they look like.  If the user ever want, say "1" to mean the string
"1" then they can manually specify the dimensions as configs.