Allow serialize to return an object pre-to_s
Normally serialization should return a string, but the generic need is that the
returned object will become the serialized string when asked to become one.
This allows numbers to be preserved in formats like JSON that don't need the
"serialized" object to ever become a string.  Kinda subtle but works for
supporting both CSV and JSON.