Remove Topic#*_each/Topic#*_data method pairs
The symmetry so nicely seen in these methods was broken with the
addition of transactions.  Transaction support was baked into
Topic#write_data, but I decided it was pretty messy to add the same to
Topic#write_each (because it would break if you got an infinite 'data'
passed in). Without symmetry it seems better to have only one version of
each of these methods.

Note the _each forms of the serialization methods were chosen because
that will allow projections to iterate without making an intermediate
Array.  Less garbage.